requirements.txt
Django>=4.1.3,<4.2
djangorestframework>=3.14,<3.15
psycopg2>=2.9.5,<3.0
Pillow>=9.3,<9.4
easy-thumbnails-rest